Work that is done on these blank files:
docs :  Add Descriptions

In This project, I tried to mimic the features that is shown in design-prototype.png file.
I customized card.html.
As you see in the complete project, CSS styles are added to the styles.css 
Border around the animal name, image and it's information (interesting fact, list items, and description) were added. I set the border-style, border-width and border-color as you can see in the completed project.
Spacing / Padding also applied around the animal's name, image, the list of items and the information.